6.1.0 - cleanup overly cautious [weak self]
Overcomplicated the API due to requiring `WaveformAnalyzer` to be retained by caller.
Actually we don't need to do that and are not creating a retain cycle.
I was just overly careful.
We don't store the analysis closure anywhere, so we're good to go.

In theory this should be 7.0 but since it's just been a day we'll just keep it at 6.1 🤫