Reset lastOffset when new drawing begins
If you reuse the Drawer between recordings, lastOffset may still contain
the value from the last recording and since it's a cumulative value
it often messes up subsequent uses.

This resets it automatically upon the new use of the drawer, making
the assumption that the only time newSampleCount and the number of
samples is the same is upon your first call.