Fix a blip when doing striped drawing
The previous code assumed that the first time that
`samples.count > samplesNeeded` was true, that `newSamplesCount`
would be the number of new samples greater than samples.count.

But this assumption does not always hold.

For example, assume that samplesNeeded is 9, and stripeBucket is 4

Then we call `draw` with the following sets of samples:

samples: [0,1,2,3,4,5,6,7] newSampleCount: 8
samples: [0,1,2,3,4,5,6,7,8,9] newSampleCount: 2

On that second call, the clipped samples is going to be:

[1,2,3,4,5,6,7,8,9]

`lastOffset` is going to be (0 + 2) % 4 = 2. But it should actually
be 1.

What this means is that the first time that `samples.count > samplesNeeded`
is true, the drawing window will pick different sample values to draw.
This may appear as a slight change in graph amplitudes when that frame
is drawn, especially when there is great variation between neighboring
samples.