render SwiftUI using custom WaveformShape
The interface could arguably be a little simpler:

- `Style` should be reduced to `.solid` and `.striped`
- the actual styling could then be simplified to solely rely on SwiftUI Shape modifiers

However, while we still allow the creation of `UIImage` / `NSImage` as well as support `UIKit`, we can't really simplify this without too much effort in the legacy (`UIKit` code),
which feels like a waste of time.

Instead, probably UIKit support should be dropped at some point in the future and the interface be overhauled for modern iOS.

Once released, this would allow #78 to be done more natively.