add back ability to position waveform vertically when using image API
This is kind of rowing back on 0447737. Intentionally not documented though to avoid the previous confusion
by not promoting its use too much and still leaving it out from the typical use cases inside views.

While views can be moved, the plain image can not, so to avoid the need of further processing, the image
API supports positioning again.